refactor : singleton_meta_class.py 생성. 싱글톤은 여기서 만들면 됨.
